+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 0.0.27
+
+- Arreglado error que trataba identificadores en una nueva linea como parametros de una función.
+- Eliminado soporte para los operadores unario ++ y --.
+
 # 0.0.23
 
 - Ahora los tokens de tipo ENumero contienen un string en lugar de un number,